Error Fixes in Battle System
Issues had with the Battle System script on Monday have been resolved.
On line 222 I added a return to make it stop yelling.  IEnumerator's need to return something whether it is null or a float, so I added a placeholder one in a random spot in PlayerAttack until we are able to add the needed return in it.